Deleting a parent element recursively deletes its children. Accordingly, the parent element's populateRequest method must report that locks are needed on the child elements. This allows the ElementEditor helper class to acquire all needed locks before deleting an assembly.

If an app were to specify both the parent element and the child elements in a single call to delete, the delete logic may find that one of the children has already been deleted (as a consequence of deleting the parent). It should just skip an already-deleted element.

ElementEditor was failing to reserve codes on new elements before inserting them.

+++ core/backend/src/IModelDb.ts
@@ -1513,7 +1513,9 @@ export namespace IModelDb {
     public deleteElement(ids: Id64Arg): void {
       const iModel = this._iModel;
       Id64.toIdSet(ids).forEach((id) => {
-        const props = this.getElementProps(id);
+        const props = this.tryGetElementProps(id);
+        if (props === undefined) // this may be a child element which was deleted earlier as a consequence of deleting its parent.
+          return;
         const jsClass = iModel.getJsClass<typeof Element>(props.classFullName) as any; // "as any" so we can call the protected methods
         jsClass.onDelete(props, iModel);
